BoardProject-main/requirements.txt
aioitertools==0.12.0
annotated-types==0.7.0
asgiref==3.8.1
asteval==1.0.6
attrs==25.2.0
bcrypt==4.3.0
boto3==1.37.11
botocore==1.37.11
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dill==0.3.9
Django==5.1.2
dnspython==2.7.0
emmet-core==0.84.6rc5
fabio==2024.9.0
fonttools==4.56.0
h5py==3.13.0
hdf5plugin==5.0.0
idna==3.10
imageio==2.37.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
larixite==2024.10.0
latexcodec==3.0.0
lazy_loader==0.4
lmfit==1.3.3
lxml==5.3.1
maggma==0.71.5
matplotlib==3.10.1
mongomock==4.3.0
monty==2025.3.3
mp-api==0.45.3
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.30.0
networkx==3.4.2
numdifftools==0.9.41
numexpr==2.10.2
numpy==1.26.4
orjson==3.10.15
packaging==24.2
palettable==3.3.3
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
psutil==7.0.0
pybtex==0.24.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydash==8.0.5
pyFAI==2025.1.0
pymatgen==2025.1.9
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pyshortcuts==1.9.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentinels==1.0.0
silx==2.2.1
six==1.17.0
smart-open==7.1.0
spglib==2.6.0
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
sshtunnel==0.4.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uncertainties==3.2.2
urllib3==2.3.0
wrapt==1.17.2
xraydb==4.5.6
xraylarch==2025.1.1
